feat: update AlbumViewModel
This commit updates the `AlbumViewModel` to correctly handle shuffle state when playing tracks or starting playback of all tracks.

When `playTrack` is called:
- If shuffle is enabled, `simpleMediaServiceHandler.setShuffledQueue` is used to set up the shuffled queue.
- If shuffle is disabled, the standard `setQueueData` and `loadMediaItem` are called.

A new `playAll` function is introduced:
- It sets the queue with all tracks from the album.
- It explicitly disables shuffle mode in `simpleMediaServiceHandler`.

The `shuffle` function is updated to use `simpleMediaServiceHandler.setShuffledQueue`.

In `AlbumScreen.kt`, the play button now calls the new `playAll` function in the view model.